% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/al_dente.R
\name{al_dente}
\alias{al_dente}
\title{Plot Dichotomous Rasch Model Residuals}
\usage{
al_dente(eRm.obj, style = "present")
}
\arguments{
\item{eRm.obj}{an eRm object of class 'RM'.}

\item{style}{Options for plot colour; a character string for plotting style
options are present for coloured plots, or print for black and white
plots. Defaults to "present".}
}
\value{
A list object containing multiple residual plots and residuals
    Pearson correlations matrix.
}
\description{
Plot Dichotomous Rasch Model Residuals
}
