% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FakeData-data.R
\docType{data}
\name{FakeData}
\alias{FakeData}
\title{Fake Data}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://www.rasch.org/rmt/rmt213a.htm}{Rasch Measurement Transactions}
}
\usage{
data(FakeData)
}
\description{
Simulated scored data for 23 items and 1200 participants, where items where presented in random order to each participant.
Contains responses for dichotomous (X), and polychotomous (K) data. Item delivery order is stored in the Index column
Data were simulated according to suggestions in Linacre (2007).
Data are stored in long format.
}
\examples{
# Show fake data
data(FakeData)

library(tidyverse)
# Count of item responses by item and polychotomous category
count_by_reponse_K <- FakeData \%>\% group_by(Item, K) \%>\%  summarise(Count = n())
}
\references{
Linacre. (2013)  Rasch Measurement Transactions 21:3 p. 1125
(\href{https://www.rasch.org/rmt/rmt213a.htm}{Rasch Measurement Transactions})
}
\keyword{datasets}
