% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FakePCMItems-data.R
\docType{data}
\name{FakePCMItems}
\alias{FakePCMItems}
\title{Fake Items}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://www.rasch.org/rmt/rmt213a.htm}{Rasch Measurement Transactions}
}
\usage{
data("FakePCMItems")
}
\description{
Item parameters for simulated scored data for 11 items and 1200 participants,
    where items where presented in random order to each participant.
    Contains responses for dichotomous (X), and polychotomous (K) data.
    Item delivery order is stored in the Index column.
    Data were simulated according to suggestions in Linacre (2007).
    These are the resulting item parameters.
}
\examples{
# Show fake item parameters
data("FakePCMItems")
}
\references{
Linacre. (2013)  Rasch Measurement Transactions 21:3 p. 1125
(\href{https://www.rasch.org/rmt/rmt213a.htm}{Rasch Measurement Transactions})
}
\keyword{datasets}
