% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_fit_table.R
\name{item_fit_table}
\alias{item_fit_table}
\title{Create Item Fit Table}
\usage{
item_fit_table(eRm.obj)
}
\arguments{
\item{eRm.obj}{input data, generated using a Rasch model estimation function
from eRm.}
}
\value{
A data.frame of class ItemFit.
}
\description{
This function extracts Item fit statistics from eRm `itemfit()`
}
\examples{
library(eRm)
library(psketti)

data("FakeData") # load data
# restructure fake data
Fake_Data_scores <- reshape(FakeData[, c("ID", "Item", "X")],
                            timevar = "Item",
                            idvar = "ID",
                            direction = "wide")
# for eRm col names and row names
names(Fake_Data_scores) <- c("ID",
                             paste0("i",
                                    sprintf(fmt  = "\%02d", 1:23)))
                                    
row.names(Fake_Data_scores) <- Fake_Data_scores$ID
Fake_Data_scores$ID         <- NULL

fake_rm     <- RM(Fake_Data_scores)    # Estimate Rasch model

itemFit_psk <- item_fit_table(fake_rm) # item fit stats

itemFit_psk # output
}
