% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psketti_msq.R
\name{pskett_msq}
\alias{pskett_msq}
\alias{psketti_msq}
\title{Plot a Infit and Outfit MSQ}
\usage{
psketti_msq(x, style = "present")
}
\arguments{
\item{x}{input data, generated using `item_fit_table()`.}

\item{style}{a character string for plotting style options are present for
coloured, or print for black and white. Defaults to "present".}
}
\value{
MSQ plot.
}
\description{
This function plots the Infit and Outfit MSQ values from a item
    of class ItemFit. see `item_fit_table()`.
}
