% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psketto.R
\name{psketto}
\alias{psketto}
\title{Plot a single ICC}
\usage{
\code{psketto(pskettified_data = df, item = "ItemName", item.label = "ItemLabel")}
}
\arguments{
\item{pskettified_data}{input data, generated using \code{pskettify()}}

\item{item}{character name of the item to be plotted}

\item{item.label}{a character string of the item name prefix.}

\item{style}{a character string for plotting style options are present for coloured plots,
or print for black and white plots. Defaults to "present".}

\item{IRFLocation}{logical, plots reference lines for Rasch IRF location. Defaults to TRUE.}

\item{empCI}{logical, plots confidence intervals for empirical points, calculated using \code{pskettify()}.
Defaults to TRUE.}

\item{empICC}{logical, plots empirical ICC for item. Defaults to \code{TRUE}.}

\item{empPoints}{logical, plots empirical points for based on class intervals/score groups
generated with \code{pskettify()}. Defaults to \code{TRUE}}
}
\value{
psketto plot.
}
\description{
psketto is singlular of psketti; (spaghetti <--> spaghetto).
    This function plots the data for a single item from the output of pskettify()
    psketto() is also used in psketti to create multiple ICC plots
}
\examples{
library(eRm)
library(psketti)
data(FakeData)  # load dataset fake data

# Set up dataframe for eRm, long to wide with dichotmous data
Fake_Data_scores <- reshape(FakeData[, c("ID", "Item", "X")],
                            timevar = "Item",
                            idvar = "ID",
                            direction = "wide")
names(Fake_Data_scores) <- c("ID", paste0("i", sprintf(fmt  = "\%02d", 1:23)))
row.names(Fake_Data_scores) <- Fake_Data_scores$ID
Fake_Data_scores$ID <- NULL

fake_rm <- RM(Fake_Data_scores) # Fit Rasch Model

psk_data <-pskettify(eRm.obj = fake_rm)

# plot IRF in default colours
psk_1_present <- psketto(psk.data, style = "present", item = 1, item.label = "Item 1")
psk_1_present

# plot IRF in default greyscale colours
psk_1_print   <- psketto(psk.data, style = "print", item = 1, item.label = "Item 1")
psk_1_print
}
