% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psketto_simple.R
\name{psketto_simple}
\alias{psketto_simple}
\title{Unadorned Rasch IRF}
\usage{
psketto_simple(x, item, all.item = FALSE, item.label = FALSE)
}
\arguments{
\item{x}{an object of class pskettified}

\item{item}{If you want to plot a single IRF, use this argument to state the name of the item.}

\item{all.item}{Should all item IRF be plotted? Defaults to FALSE}

\item{item.label}{Should the item labels be plotted? Defaults to FALSE}
}
\description{
Unadorned Rasch IRF
}
\examples{
library(eRm)
library(psketti)

data("FakeData") # load data
# restructure fake data
Fake_Data_scores <- reshape(FakeData[, c("ID", "Item", "X")],
                            timevar = "Item",
                            idvar = "ID",
                            direction = "wide")
# for eRm col names and row names
names(Fake_Data_scores) <- c("ID",
                             paste0("i",
                                    sprintf(fmt  = "\%02d", 1:23)))

row.names(Fake_Data_scores) <- Fake_Data_scores$ID
Fake_Data_scores$ID         <- NULL

fake_rm   <- RM(Fake_Data_scores) # Estimate Rasch model

psk_data  <- pskettify(fake_rm)   # pskettify data
psketto_simple(psk_data, item = "i01")                       # single item IRF, no label
psketto_simple(psk_data, item = "i01", item.label = TRUE)    # single item IRF, labeled
psketto_simple(psk_data, item.label = TRUE, all.item = TRUE) # all item IRF labeled
}
