% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_gen.R
\name{report_gen}
\alias{report_gen}
\title{Generate an Item Report}
\usage{
report_gen(
  item.names,
  eRm.obj,
  data.long,
  Title,
  ptrn,
  ID,
  Item,
  K,
  r_o,
  html_gen = FALSE
)
}
\arguments{
\item{item.names}{A charcter list of item names}

\item{eRm.obj}{dichotomous Rasch Model; an eRm object of class 'RM'.}

\item{data.long}{Long formatted data set with a column for Item names
matching item.names, a column for response options, matching r_o, and an
ID column name matching ID.}

\item{Title}{the title of the Rmd report.}

\item{ptrn}{the prefix label of the item namesL eg i01 is "i".}

\item{ID}{A character object giving the name of the ID column in 
data.long}

\item{Item}{A character object giving the name of the Item column in 
data.long}

\item{K}{A character object giving the name of the response column in 
data.long}

\item{r_o}{An ordered factor object to arrange columns along by response
option order.}

\item{html_gen}{Generate a html file from the report? Defaults to FALSE.
Currently not functional.}
}
\value{
A folder containing materials for report generation

If html_gen = TRUE, a html report will be rendered
}
\description{
report_gen generates a containing folder for an Rmarkdown file, in the root
    of the project directory. After report_gen has been run, you can go to
    the containing folder and render the Rmarkdown file. Future versions will
    allow you to generate the Rmd file and render the html file in one go by
    selecting html_gen = TRUE.  However, you will need to have tidyverse,
    knitr rmarkdown and kableExtra installed. This function is still
    undergoing active development.
}
