% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabliatelle.R
\name{tabliatelle}
\alias{tabliatelle}
\title{Distractor Analysis Tables}
\usage{
tabliatelle(x, ID, Item, K, response_options, eRm.obj)
}
\arguments{
\item{x}{A long formatted dataframe.}

\item{ID}{column name for ID column.}

\item{Item}{column name for Item column.}

\item{K}{column name for column containing multiple choice responses.}

\item{response_options}{An ordered factor object to arrange column order in
the distractor table.}

\item{eRm.obj}{An object of class eRm and model RM. Use `eRm::RM(score_data)`
to create this object.}
}
\value{
tabliatelle returns a list of class tabliatelle.
}
\description{
Implementation of the tabular (Andrich and Styles, 2009)
    approach to assigning a partial credit scoring system to data previously
    modeled with a dichotomous Rasch model.
}
\examples{
library(eRm)
library(psketti)

data("FakeData") # load data
# restructure fake data
Fake_Data_scores <- reshape(FakeData[, c("ID", "Item", "X")],
                            timevar = "Item",
                            idvar = "ID",
                            direction = "wide")
# for eRm col names and row names
names(Fake_Data_scores) <- c("ID",
                             paste0("i",
                                    sprintf(fmt  = "\%02d", 1:23)))
                                    
row.names(Fake_Data_scores) <- Fake_Data_scores$ID
Fake_Data_scores$ID         <- NULL

fake_rm   <- RM(Fake_Data_scores) # Estimate Rasch model

# Prepare response options factor
r_o <- factor(sort(unique(FakeData$K)),          # input var
              levels = sort(unique(FakeData$K)), # factor levels
              ordered = TRUE)                    # ordered
# tabliatellify
tlt_data <- tabliatelle(x = FakeData,
                        eRm.obj = fake_rm,
                        ID = "ID",
                        Item = "Item",
                        K = "K",
                        response_options = r_o)

tlt_data # output
}
