% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingrediente.R
\name{ingrediente}
\alias{ingrediente}
\title{Score Report Tables}
\usage{
\code{ingrediente(x = df, ID = "ID", Item = "Item", K = "K", K_options = r_o, Index = "Index")}
}
\arguments{
\item{ID}{column name for ID column}

\item{Item}{column name for Item column}

\item{Score}{a column name for response scores}

\item{K}{column name for column containing multiple choice responses}

\item{K_options}{An ordered factor object to arrange column order in the score table.}

\item{Index}{Column name for order of administration per participant.
This can also be an ordered factor for the item names.
Orders the Response string. Defaults to \code{NULL}, using the items to order the response string.}
}
\value{
Score reports for participants, with counts of category selection and a score string ordered by score string index
}
\description{
Generates score report for dataframe.
}
\examples{
data(FakeData)

K_opt <- factor(LETTERS[1:5], levels = LETTERS[1:5], ordered = TRUE)
score_report <- ingrediente(x = FakeData,
                            Item = "Item",
                            ID = "ID",
                            Score = "X",
                            K = "K",
                            K_options = K_opt,
                            Index = "Index")

# show score report for values with a total score <= 5
score_report[score_report$total_score <= 1, ]

# Score report ordering response string by item difficulty
data("FakeItems")
FI2 <- FakeItems[order(FakeItems$Beta),]
row.names(FI2)<- NULL
FI_factor <- factor(FI2$Item, levels = FI2$Item, ordered = TRUE)



score_report2 <- ingrediente(x = FakeData,
                             Item = "Item",
                             ID = "ID",
                             Score = "X",
                             K = "K",
                             K_options = K_opt,
                             Index = FI_factor)

# show score report for values with a total score <= 5
score_report2[score_report2$total_score == 21, ]
}
