% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_fit_table.R
\name{item_fit_table}
\alias{item_fit_table}
\title{Create Item Fit Table}
\usage{
item_fit_table(eRm.obj)
}
\arguments{
\item{eRm.obj}{input data, generated using a Rasch model estimation function
from eRm.}
}
\value{
A data.frame of class ItemFit.
}
\description{
This function extracts Item fit statistics from eRm `itemfit()`
}
