% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psketti_msq.R
\name{pskett_msq}
\alias{pskett_msq}
\alias{psketti_msq}
\title{Plot a Infit and Outfit MSQ}
\usage{
psketti_msq(x, style = "present")
}
\arguments{
\item{x}{input data, generated using `item_fit_table()`.}

\item{style}{a character string for plotting style options are present for
coloured, or print for black and white. Defaults to "present".}
}
\value{
MSQ plot.
}
\description{
This function plots the Infit and Outfit MSQ values from a item
    of class ItemFit. see `item_fit_table()`. Based on Yu (2020).
}
\examples{
library(eRm)
library(psketti)

data("FakeData") # load data
# restructure fake data
Fake_Data_scores <- reshape(FakeData[, c("ID", "Item", "X")],
                            timevar = "Item",
                            idvar = "ID",
                            direction = "wide")
# for eRm col names and row names
names(Fake_Data_scores) <- c("ID",
                             paste0("i",
                                    sprintf(fmt  = "\%02d", 1:23)))
                                    
row.names(Fake_Data_scores) <- Fake_Data_scores$ID
Fake_Data_scores$ID         <- NULL

fake_rm     <- RM(Fake_Data_scores)     # Estimate Rasch model

itemFit_psk <- item_fit_table(fake_rm)  # item fit stats
MSQplot     <- psketti_msq(itemFit_psk) # Plot infit and outfit

MSQplot
}
